+++ config/urls.py
@@ -8,6 +8,7 @@ urlpatterns = [
     path("api/facts/", include("facts.urls")),
     path("api/webconfig/", include("webconfig.urls")),
     path("api/processes/", include("processes.urls")),
+    path("api/was/", include("was.urls")),
     path("api/", include("core.urls")),
     path("api/", include(dashboard_api_urlpatterns)),
     path("dashboard/", include("dashboard.urls")),